Revert "CRN-1213 extract TabbedContent for working group double card (#2360)"
This reverts commit 8e58f4fbcfcdc527b84c24a22d257968e2592579.
