+++ packages/react-components/src/molecules/TabbedCard.tsx
@@ -1,10 +1,26 @@
 import { css } from '@emotion/react';
-import { ReactNode } from 'react';
-import { Card } from '../atoms';
-import { TabbedContent } from '.';
+import React, { ReactNode, useEffect } from 'react';
+import { Card, Paragraph, Headline3, TabButton, Button } from '../atoms';
+import { steel } from '../colors';
+import { perRem } from '../pixels';
+import { TabNav } from '.';
+import { paddingStyles } from '../card';
 
 const cardStyles = css({
   overflow: 'hidden',
+});
+
+const headerStyles = css({
+  paddingBottom: 0,
+  borderBottom: `1px solid ${steel.rgb}`,
+});
+
+const showMoreStyles = css({
+  display: 'flex',
+  justifyContent: 'center',
+  paddingTop: `${16 / perRem}em`,
+  paddingBottom: `${16 / perRem}em`,
+  borderTop: `1px solid ${steel.rgb}`,
 });
 
 export type TabProps<T> = {
@@ -30,18 +46,51 @@ const TabbedCard = <T extends object>({
   activeTabIndex = 0,
   getShowMoreText,
   children,
-}: TabbedCardProps<T>) => (
-  <Card padding={false} overrideStyles={cardStyles}>
-    <TabbedContent
-      title={title}
-      description={description}
-      tabs={tabs}
-      activeTabIndex={activeTabIndex}
-      getShowMoreText={getShowMoreText}
-    >
-      {children}
-    </TabbedContent>
-  </Card>
-);
+}: TabbedCardProps<T>) => {
+  useEffect(() => setActive(activeTabIndex), [activeTabIndex]);
+  const [active, setActive] = React.useState(activeTabIndex);
+  const [showMore, setShowMore] = React.useState(false);
+  const { items, truncateFrom } = tabs[active];
+  const displayShowMoreButton =
+    getShowMoreText && truncateFrom && items.length > truncateFrom;
+
+  return (
+    <Card padding={false} overrideStyles={cardStyles}>
+      <div css={[paddingStyles, headerStyles]}>
+        <Headline3>{title}</Headline3>
+        <Paragraph hasMargin={false} accent="lead">
+          {description}
+        </Paragraph>
+        <TabNav>
+          {tabs.map(({ tabTitle, disabled }, index) => (
+            <TabButton
+              key={tabTitle}
+              active={index === active}
+              onClick={() => {
+                setShowMore(false);
+                setActive(index);
+              }}
+              disabled={disabled}
+            >
+              {tabTitle}
+            </TabButton>
+          ))}
+        </TabNav>
+      </div>
+      <div css={[paddingStyles, { paddingBottom: 0, paddingTop: 0 }]}>
+        {children({
+          data: items.slice(0, showMore ? undefined : truncateFrom),
+        })}
+      </div>
+      {displayShowMoreButton && (
+        <div css={showMoreStyles}>
+          <Button linkStyle onClick={() => setShowMore(!showMore)}>
+            {getShowMoreText(showMore)}
+          </Button>
+        </div>
+      )}
+    </Card>
+  );
+};
 
 export default TabbedCard;
