+++ packages/react-components/src/molecules/index.ts
@@ -50,4 +50,3 @@ export { default as UserMenuButton } from './UserMenuButton';
 export { default as UserProfilePersonalText } from './UserProfilePersonalText';
 export { default as UsersList } from './UsersList';
 export { default as UtilityBar } from './UtilityBar';
-export { default as TabbedContent } from './TabbedContent';
